requirements.txt
smolagents[all]
transformers>=4.37.0
helium>=3.0.0
selenium>=4.0.0
python-dotenv>=0.19.0
Pillow>=10.1.0
click>=8.1.7
pandas>=1.5.0
openpyxl>=3.0.0
undetected-chromedriver>=3.5.5,<4.0.0
torch>=2.1.0
fireworks-ai>=0.6.0  # For vision model access